Enable shetho at only debug
