+++ app/src/debug/AndroidManifest.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    package="seki.com.doyouworkout">
+
+    <application
+        tools:replace="android:name"
+        android:name=".DebugApp"
+        tools:ignore="GoogleAppIndexingWarning">
+    </application>
+
+</manifest>
