+++ app/src/debug/java/seki/com/doyouworkout/DebugApp.kt
@@ -1,0 +1,11 @@
+package seki.com.doyouworkout
+
+import com.facebook.stetho.Stetho
+
+class DebugApp : BaseApp() {
+
+    override fun onCreate() {
+        super.onCreate()
+        Stetho.initializeWithDefaults(this)
+    }
+}
