+++ app/src/main/java/seki/com/doyouworkout/App.kt
@@ -1,21 +1,3 @@
 package seki.com.doyouworkout
 
-import com.facebook.stetho.Stetho
-import dagger.android.AndroidInjector
-import dagger.android.DaggerApplication
-import seki.com.doyouworkout.di.ApplicationModule
-import seki.com.doyouworkout.di.DaggerApplicationComponent
-
-class App: DaggerApplication() {
-
-    override fun onCreate() {
-        super.onCreate()
-        Stetho.initializeWithDefaults(this)
-    }
-
-    override fun applicationInjector(): AndroidInjector<out DaggerApplication> {
-        return DaggerApplicationComponent.builder()
-                .applicationModule(ApplicationModule(this))
-                .build()
-    }
-}
+class App: BaseApp()
